Extract rust_stream inline test suite to src/tests.rs
main.rs shrinks from 588 to 154 lines; the mock-antd test suite moves
verbatim to its own file.
